Moved Period entity into the CourseMembership subsystem
