+++ app/subsystems/course_membership/period.rb
@@ -1,0 +1,13 @@
+module CourseMembership
+  class Period < Entity
+
+    wraps CourseMembership::Models::Period
+
+    exposes :course, :name, :student_roles, :teacher_roles
+
+    def to_model
+      repository
+    end
+
+  end
+end
