+++ spec/representers/api/v1/student_representer_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe Api::V1::StudentRepresenter, type: :representer do
   let!(:user) { FactoryGirl.create(:user_profile).entity_user }
-  let!(:period) { FactoryGirl.create(:period) }
+  let!(:period) { FactoryGirl.create(:course_membership_period) }
   let!(:student) {
     AddUserAsPeriodStudent.call(period: period, user: user).outputs.student
   }
